+++ liveness-server-typescript/src/common.ts
@@ -1,0 +1,6 @@
+
+export enum LivenessMode {
+    DetectLiveness = 'detectLiveness',
+    DetectLivenessWithVerify = 'detectLivenessWithVerify'
+  }
+
